Removing the need for connector to pass DatastreamTask while creating the DatastreamEventRecord
